Change testing strategy
Most of those tests I wrote over the last 10ish commits suck. I was
testing methods, not implementation. The big red flag I should have seen
was stubbing methods on the tested class!

I'm not going to test private methods, or methods in general. I'm going
to test each class as a whole, against all the various inputs it will
receive. The 'unit' in unit test isn't a method, it's a class.

This commit eliminates tests of classes that are never instantiated, and
tests the img output format. All other PictureTag classes are stubbed,
but ObjectiveElements is left untouched since it is so integral to its
output. I could remove this dependency by stubbing the OE API, but it's
a tiny and simple library and simply verifying an output string is much
easier.