Refactor source_image hash into an array
This fixes one of my """clever""" ideas.

Old way: Source images are tracked in a hash. Keys are media query
names, meaning that source images are referred to by their media
queries. The first, most important image has no media query, so its key
is nil. They are given in order of increasing specificity.

This relied on two quirks of Ruby hashes: that nil is a perfectly fine
key, and hashes can be trusted to keep their order. While those things
are true, people don't normally think of hashes as having an order, nor
is it very intuitive to use nil as key, much less for the primary source
image, nor are media queries the most important thing about source
images.

New way: Source images are an array, in order of increasing specificity.
Their associated media queries are instance methods, and the primary
source image is just the first one in the array. The end result is
pretty much the same, but now hopefully I won't have a WTF moment every
time I come back to this.