Combine heavy image generation steps
Previously we performed image resizing, reformatting, and quality
adjustment individually. This caused some problems with generated image
quality and hurt performance as the image had to be written repeatedly.
Now, all 3 arguments will be passed to ImageMagick, which gives it the
freedom to be smart about things and produces better results overall.