Disable Minitest/TestMethodName cop
Our tests aren't quite set up the way the cop wants them to be. Many of
them define a 'tested' method, referencing whatever object is the
subject of our test. The cop seems to think this is just another test
and wants them to be named test_tested.

I'm not sure why this cop exists anyway; minitest only runs methods
whose name start with 'test_'. If it doesn't start with 'test_', it's
not a test. Perhaps it's a hint that our tests need some refactoring,
which is valid.