Update crop-related settings
* Rename 'geometry' to 'crop'
  * This will now only accept an aspect ratio in the 'w:h' format.
* Rename 'gravity' to 'keep'
  * Valid Options are 'center', 'entropy', and 'attention'.
* Add list of valid 'keep' settings
* Remove now unnecessary constants

Crop handling will be the biggest change to user-facing syntax in this
update. The old settings were thin wrappers around Imagemagick's crop
syntax, which accepts a geometry and gravity. The new settings will
accept an aspect ratio, and an instruction on which part of the image to
keep.

JPT is not an image editor. If the existing smartcrop feature doesn't
produce satisfactory results, fire up a proper editor.