Add format-specific write-options
libvips by default uses the wrong compression method for avif files,
resulting in image files that chrome & others cannot read. We have to
specify av1 compression for it to work, and if we're doing that we might
as well add support for other format-specific write options.

This functionality could swallow up the image quality setting, and if I
were writing this from scratch I would consider it, but there's enough
momentum behind the current implementation that I'm not going to change
it. Also, that method wouldn't support the size-dependent quality
feature we've created as nicely.