Implement a global threadpool for faster builds.
Fixes #225

This implements a global Concurrent::ThreadPoolExecutor from
concurrent-ruby to avoid memory blowup, and de-duplicates files before
generation so we no longer need to rely on filesystem consistency to
avoid double-generating images.

This is an alternate to #282 with a litle bit more complexity, but with
the added benefits that all image generation can happen in a single
ThreadPool.